refactor: menu component
Refactored MainMenu into MainMenuWrapper and introduced Menu as a component that we can use both in footer and header.
