--- src/components/MainMenu/index.ts
@@ -1,3 +1,0 @@
-import MainMenu from './MainMenu'
-
-export default MainMenu
