+++ src/components/MainMenuWrapper/index.ts
@@ -1,0 +1,3 @@
+import MainMenuWrapper from './MainMenuWrapper'
+
+export default MainMenuWrapper
